seperated cli and functions
